rename: RosterPage를 PlayerPage로 변경
