+++ src/pages/club/PlayerPage.tsx
